Within Group Distribution Calculation
This was a really fun piece of code to write- it involves a function
that identifies the correct comparison group and uses the ecdf R
function to produce the individual score's position on the group
distribution. We used it for student achievement among peers but it
could be useful for a lot of things.